requirements.txt
autopep8==1.5.4
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coloredlogs==14.0
datefinder==0.7.1
dateparser==0.7.6
decorator==4.4.2
dnspython==1.16.0
Flask==1.1.2
flask-session==0.3.2
google-api-core==1.22.2
google-api-python-client==1.11.0
google-auth==1.21.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
humanfriendly==8.2
idna==2.10
IMAPClient==2.1.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
Pygments==2.6.1
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tzlocal==2.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
